+++ resources/lib/UnityOrg.php
@@ -11,14 +11,16 @@ class UnityOrg
     private $MAILER;
     private $SQL;
     private $LDAP;
+    private $REDIS;
 
-    public function __construct($orgid, $LDAP, $SQL, $MAILER)
+    public function __construct($orgid, $LDAP, $SQL, $MAILER, $REDIS)
     {
         $this->orgid = $orgid;
 
         $this->LDAP = $LDAP;
         $this->SQL = $SQL;
         $this->MAILER = $MAILER;
+        $this->REDIS = $REDIS;
     }
 
     public function init()
@@ -35,6 +37,8 @@ class UnityOrg
                 throw new Exception("Failed to create POSIX group for " . $this->orgid);  // this shouldn't execute
             }
         }
+
+        $this->REDIS->appendCacheArray("sorted_orgs", "", $this->getOrgID());
     }
 
     public function exists()
@@ -47,11 +51,48 @@ class UnityOrg
         return $this->LDAP->getOrgGroupEntry($this->orgid);
     }
 
+    public function getOrgID()
+    {
+        return $this->orgid;
+    }
+
     public function inOrg($user)
     {
         $org_group = $this->getLDAPOrgGroup();
         $members = $org_group->getAttribute("memberuid");
         return in_array($user, $members);
+    }
+
+    public function getOrgMembers($ignorecache = false)
+    {
+        if (!$ignorecache) {
+            $cached_val = $this->REDIS->getCache($this->getOrgID(), "members");
+            if (!is_null($cached_val)) {
+                $members = $cached_val;
+            }
+        }
+
+        $updatecache = false;
+        if (!isset($members)) {
+            $org_group = $this->getLDAPOrgGroup();
+            $members = $org_group->getAttribute("memberuid");
+            $updatecache = true;
+        }
+
+        $out = array();
+        $cache_arr = array();
+        foreach ($members as $member) {
+            $user_obj = new UnityUser($member, $this->LDAP, $this->SQL, $this->MAILER, $this->REDIS);
+            array_push($out, $user_obj);
+            array_push($cache_arr, $user_obj->getUID());
+        }
+
+        if (!$ignorecache && $updatecache) {
+            sort($cache_arr);
+            $this->REDIS->setCache($this->getOrgID(), "members", $cache_arr);
+        }
+
+        return $out;
     }
 
     public function addUser($user)
@@ -62,6 +103,8 @@ class UnityOrg
         if (!$org_group->write()) {
             throw new Exception("Unable to write to org group");
         }
+
+        $this->REDIS->appendCacheArray($this->getOrgID(), "members", $user->getUID());
     }
 
     public function removeUser($user)
@@ -72,5 +115,7 @@ class UnityOrg
         if (!$org_group->write()) {
             throw new Exception("Unable to write to org group");
         }
+
+        $this->REDIS->removeCacheArray($this->getOrgID(), "members", $user->getUID());
     }
 }