+++ webroot/index.php
@@ -7,7 +7,7 @@ require_once $LOC_HEADER;
 
 
 <?php
-echo $SQL->getPage($BRANDING["page"]["home"])["content"];
+echo $SQL->getPage($CONFIG["page"]["home"])["content"];
 ?>
 
 <h1>Cluster Notices</h1>