+++ webroot/panel/modal/pi_search.php
@@ -7,7 +7,7 @@ if (empty($search_query)) {
     die("<span>No Results</span>");
 }
 
-$assocs = $LDAP->getAllPIGroups($SQL, $MAILER);
+$assocs = $LDAP->getAllPIGroups($SQL, $MAILER, $REDIS);
 
 $MAX_COUNT = 10;  // Max results of PI search